+++ src/main/java/com/tensai/cms/telegram/api/events/CreateTopicEvent.java
@@ -11,7 +11,7 @@ public record CreateTopicEvent(
         Long chatId,
 
         @JsonProperty("message_thread_id")
-        Integer messageThreadId,
+        Long messageThreadId,
 
         @JsonProperty("topic_name")
         String topicName
